backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
requests==2.31.0
boto3==1.28.57
python-dotenv==1.0.0
prometheus-client==0.17.1
schedule==1.2.0
psutil==5.9.5
pandas==2.0.3
numpy==1.24.3
colorlog==6.8.0
